+++ .coveragerc
@@ -1,7 +1,7 @@
 # .coveragerc to control coverage.py
 [run]
 branch = True
-source = imgdataconvertcodegen
+source = im2im
 # omit = bad_file.py
 
 [paths]
