+++ setup.cfg
@@ -4,7 +4,7 @@
 # https://setuptools.pypa.io/en/latest/references/keywords.html
 
 [metadata]
-name = ImgDataConvertCodeGen
+name = im2im
 description = Automatically generating conversion code for in-memory representations of images using a knowledge graph of data types.
 author = Fei Chen
 author_email = boxchenfei@gmail.com
@@ -12,12 +12,12 @@ license = MIT
 license_files = LICENSE.txt
 long_description = file: README.md
 long_description_content_type = text/markdown; charset=UTF-8
-url = https://github.com/c3di/ImgDataConvertCodeGen
+url = https://github.com/c3di/im2im
 # Add here related links, for example:
 project_urls =
-;     Documentation = https://github.com/c3di/ImgDataConvertCodeGen
-    Source = https://github.com/c3di/ImgDataConvertCodeGen/
-    Tracker = https://github.com/c3di/ImgDataConvertCodeGen/issues
+;     Documentation = https://github.com/c3di/im2im
+    Source = https://github.com/c3di/im2im/
+    Tracker = https://github.com/c3di/im2im/issues
 #    Download = https://pypi.org/project/PyScaffold/#files
 
 # Change if running only on Windows, Mac or Linux (comma-separated)
@@ -60,7 +60,7 @@ exclude =
 
 [options.extras_require]
 # Add here additional requirements for extra features, to install with:
-# `pip install ImgDataConvertCodeGen[PDF]` like:
+# `pip install im2im[PDF]` like:
 # PDF = ReportLab; RXP
 
 # Add here test requirements (semicolon/line-separated)
@@ -72,7 +72,7 @@ testing =
 [options.entry_points]
 # Add here console scripts like:
 console_scripts =
-    ImgDataConvertCodeGen = imgdataconvertcodegen.interface_cmd:run
+    im2im = im2im.interface_cmd:run
 
 [tool:pytest]
 # Specify command line options as you would do when invoking pytest directly.
@@ -81,7 +81,7 @@ console_scripts =
 # CAUTION: --cov flags may prohibit setting breakpoints while debugging.
 #          Comment those flags to avoid this pytest issue.
 addopts =
-    --cov imgdataconvertcodegen --cov-report term-missing
+    --cov im2im --cov-report term-missing
     --verbose
 norecursedirs =
     dist
@@ -116,4 +116,4 @@ exclude =
 # PyScaffold's parameters when the project was created.
 # This will be used when updating. Do not change!
 version = 4.5
-package = imgdataconvertcodegen
+package = im2im
