+++ src/im2im/knowledge_graph_construction/edge_factories/numpy.py
