+++ tests/test_metadata.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.imgdataconvertcodegen.knowledge_graph_construction.metedata.util import find_closest_metadata, encode_metadata, \
+from src.im2im.knowledge_graph_construction.metedata.util import find_closest_metadata, encode_metadata, \
     decode_metadata
